implementation

+++ src/main/scala/com/scalaspark/models/PlayList.scala
@@ -1,5 +1,5 @@
 package com.scalaspark.models
 
-class PlayList(id:Int, name:String) {
+class PlayList(id:Long, name:String) {
 
 }
